tags: validate tag values.
This makes sure that | and : are not sent in as tag values, which causes
statsd parse errors, since | and : are delimiters in the statsd format.

This is the most conservative approach in validation, and I think it
makes sense because people tend to send unchecked input in as tag
values. There are stricter rules for metric names and tag keys (at least
in wavefront[1]) but rather than silently changing those, we should let
them fail and cause alerts as they currently do.

[1] https://docs.wavefront.com/wavefront_data_format.html