Add thread-safe mock sink for testing (#68)
A common error in test code is to use a racy Store or Sink implementation and the existing MockSink cannot be used in a thread-safe manner by consumers of this library.  This PR adds a thread-safe Sink implementation that can be used with the existing Store code.

Additionally, the Sink has helper methods to make testing easier.