Remove the option to publish vars via expvars. (#76)
Expvars exposes the program's command line arguments at /var/debug -
this cannot be disabled and is a security risk.  It also panics and logs
if a duplicate var name is passed in (because we use it incorrectly).
Additionally, it doesn't appear that anything still relies on the
memory or other stats published by expvars - so remove it.