tests: disable running tests in parallel in Travis CI (#109)
Travis CI is slow and this makes tests that have timeouts flaky. We
could increase the timeouts, but that would slow the feedback cycle of
local development. Instead this commit disables running tests in
parallel when the 'travis' build tag is provided (I had some spicier
names for it, but I considering this is OSS 'travis' will do).

Note: we do not rely on tests being parallel to identify race conditions
- we do that explicitly and often.