Migrate to GHA (#114)
* Migrate to GHA

Also remove the Makefile. vet is replaced with `-vet=all`. We're not
running quick anymore (no point in running the tests with the race
detector disabled only to run it again with the race detector enabled).

Finally, for linting, adopt golangci-lint.

* Fix golangci-lint violations

* Add dependabot config

This will help us keep GHA and go modules up to date.

* Don't run tests twice

* Remove hack for not running tests in parallel in Travis

* readme: fix badge

Want to point the svg to the right branch too.

Co-authored-by: francisco souza <fsouza@users.noreply.github.com>