remove logrus dependency
This PR removes logrus as a dependency, and makes the built-in
default logger work more like Zap than it does Logrus.

In the future, users of gostats should bring their own
logger. Zap is fairly popular, so I expect many folks to conform to
Zap like logging, so the default logger now logs in JSON formatted
lines like zap.NewProduction, though not exactly.

If the built-in default logger is not suitable for a user's needs,
they can use a zap sugared logger, logrus, or conform to the interface
with their own logger.