don't flush zero counters (#136)
This undoes a change in 0ed02ca2404195b02e124f16685a53046bbafa06,
which always emitted counters even when if the value was set to zero,
i.e. declared, but never incremented.

The idea behind changing this back is to have some consistency - some
orgs like to skip sending zeros in order to save on network costs;
when sending them, aggregation systems tend to drop them anyway. It's
best to not emit them at all.