allow periodic flushes to Stop being emitted. (#146)
* allow periodic flushes to Stop being emitted.

This introduces a new Stop() method and changes the behavior of the
Start() method on a store just slightly.

store.Start(), if called while a periodic flush goroutine is currently
running, causes the existing goroutine to stop and be replaced with a
new goroutine. That is, the caller can now change the rate of flushing
with calls such as:

store.Start(time.NewTicker(1*time.Second))
store.Start(time.NewTicker(100*time.Second))

store.Stop() will stop any existing running periodic flush
goroutine. If called multiple times, it is a no-op.

* close instead of sending a bool

* use the values from the closure