Add reserved_words counter (#168)
We've run into an issue internally where certain applications are
emitting stats with reserved words. We don't want to panic or otherwise
harm the app, but we do want to be able to alert / notify owners when
this is happening.

We create an additional counter here to track how often reserved_tags
are being sent to statsd without blocking them.